Move nim service stuff into new nimServiceOperator flag
