+++ packages/nim-serving/package.json
@@ -12,7 +12,7 @@
     "type-check": "tsc --noEmit"
   },
   "exports": {
-    "./extensions": "./extensions.ts",
+    "./extensions": "./extensions/index.ts",
     "./types": "./src/api/nimservices/types.ts"
   },
   "dependencies": {
