+++ packages/nim-serving/src/constants.ts
@@ -1,1 +1,7 @@
 export const NIM_OPERATOR_MANAGED_BY = 'k8s-nim-operator';
+
+/** Platform id for the legacy NIM UI (KServe-backed NIM serving). */
+export const NIM_LEGACY_ID = 'nvidia-nim';
+/** Platform id for NIMService deployments managed by the k8s-nim-operator. */
+export const NIM_SERVICE_ID = 'nvidia-nim-service';
+export const NIM_MODEL_TYPE = 'NVIDIA NIM';
