+++ packages/nim-serving/src/pages/deploymentWizard/extractNIMFormData.ts
@@ -10,7 +10,7 @@ import type { ModelTypeFieldData } from '@odh-dashboard/model-serving/shared/wiz
 import type { ExtractionResult } from '@odh-dashboard/model-serving/extension-points';
 import type { NIMDeployment } from '../../api/nimservices/types';
 import { NIM_SERVICE_HARDWARE_PROFILE_PATHS } from '../../api/nimservices/utils';
-import { NIM_MODEL_TYPE } from '../../../extensions';
+import { NIM_MODEL_TYPE } from '../../constants';
 
 export { NIM_SERVICE_HARDWARE_PROFILE_PATHS };
